updated(https://github.com/floooh/sokol/commit/d40b639cbeadadffd0d6cbf74a9948e7c320d2ea)

+++ src/sokol/app.d
@@ -164,7 +164,7 @@ enum AndroidTooltype {
 }
 extern(C)
 struct Touchpoint {
-    size_t identifier;
+    ulong identifier;
     float pos_x;
     float pos_y;
     AndroidTooltype android_tooltype;
